requirements.txt
Click==7.0
Flask==1.1.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mongoengine==0.18.2
pymongo==3.8.0
python-dotenv==0.10.3
six==1.12.0
Werkzeug==0.15.5
WTForms==2.2.1
